@@ -16,35 +16,35 @@
     "require": {
         "php": "^8.1",
         "ext-json": "*",
-        "bepsvpt/secure-headers": "^7.2",
-        "doctrine/dbal": "^3.3",
+        "bepsvpt/secure-headers": "^7.3",
+        "doctrine/dbal": "^3.5",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^9.9",
-        "laravel/horizon": "^5.9",
-        "laravel/tinker": "^2.7",
-        "laravel/ui": "^3.4",
-        "livewire/livewire": "^2.0",
-        "owenvoke/blade-fontawesome": "^2.0",
+        "laravel/framework": "^10.2",
+        "laravel/horizon": "^5.12",
+        "laravel/tinker": "^2.8",
+        "laravel/ui": "^4.2",
+        "livewire/livewire": "^2.11",
+        "owenvoke/blade-fontawesome": "^2.2",
         "spatie/commonmark-highlighter": "^3.0",
         "spatie/laravel-backup": "^8.1",
-        "spatie/laravel-feed": "^4.1",
-        "spatie/laravel-honeypot": "^4.2",
+        "spatie/laravel-feed": "^4.2",
+        "spatie/laravel-honeypot": "^4.3",
         "spatie/laravel-mailcoach": "^5.0",
         "spatie/laravel-mailcoach-postmark-feedback": "^4.0",
-        "spatie/laravel-mailcoach-unlayer": "^2.0",
-        "spatie/laravel-responsecache": "^7.2",
-        "spatie/laravel-sitemap": "^6.1",
+        "spatie/laravel-mailcoach-unlayer": "^3.1",
+        "spatie/laravel-responsecache": "^7.4",
+        "spatie/laravel-sitemap": "^6.2",
         "spatie/laravel-welcome-notification": "^2.3",
         "spatie/schema-org": "^3.11",
         "spatie/yaml-front-matter": "^2.0",
         "league/flysystem-aws-s3-v3": "^3.0",
         "league/flysystem-ftp": "^3.0",
-        "symfony/postmark-mailer": "^6.0",
-        "symfony/http-client": "^6.0",
-        "spatie/laravel-ignition": "^1.2"
+        "symfony/postmark-mailer": "^6.2",
+        "symfony/http-client": "^6.2",
+        "spatie/laravel-ignition": "^2.0"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.6",
+        "barryvdh/laravel-debugbar": "^3.8",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^6.1",
         "phpunit/phpunit": "^9.5.10",
@@ -75,7 +75,7 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "post-update-cmd": [
