@@ -33,10 +33,8 @@ class CachePostsCommand extends Command
 
     /**
      * Execute the console command.
-     *
-     * @return int
      */
-    public function handle()
+    public function handle(): int
     {
         PostCollector::all();
 
