@@ -23,10 +23,8 @@ class GenerateSitemap extends Command
 
     /**
      * Execute the console command.
-     *
-     * @return mixed
      */
-    public function handle()
+    public function handle(): void
     {
         SitemapGenerator::create(config('app.url'))
             ->writeToFile(public_path('sitemap.xml'));
