@@ -10,10 +10,8 @@ class BladeServiceProvider extends ServiceProvider
 {
     /**
      * Bootstrap services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         Blade::component('layouts.main', MainLayout::class);
     }
