@@ -25,8 +25,6 @@ class RouteServiceProvider extends ServiceProvider
 
     /**
      * Define your route model bindings, pattern filters, etc.
-     *
-     * @return void
      */
     public function boot(): void
     {
@@ -35,12 +33,10 @@ class RouteServiceProvider extends ServiceProvider
 
     /**
      * Define the routes for the application.
-     *
-     * @return void
      */
     public function map(): void
     {
-        Route::group(['middleware' => ['doNotCacheResponse']], function () {
+        Route::middleware('doNotCacheResponse')->group(function () {
             Route::mailcoach('mailcoach');
             Route::mailcoachUnlayer('mailcoachUnlayer');
         });
@@ -56,8 +52,6 @@ class RouteServiceProvider extends ServiceProvider
      * Define the "web" routes for the application.
      *
      * These routes all receive session state, CSRF protection, etc.
-     *
-     * @return void
      */
     protected function mapWebRoutes(): void
     {
@@ -65,7 +59,6 @@ class RouteServiceProvider extends ServiceProvider
             ->group(base_path('routes/web.php'));
 
         Route::middleware('web')
-            ->namespace($this->namespace)
             ->group(base_path('routes/privacy.php'));
     }
 
@@ -73,8 +66,6 @@ class RouteServiceProvider extends ServiceProvider
      * Define the "api" routes for the application.
      *
      * These routes are typically stateless.
-     *
-     * @return void
      */
     protected function mapApiRoutes(): void
     {
