@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
 return new class extends Migration {
-    public function up()
+    public function up(): void
     {
         Schema::create('webhook_calls', function (Blueprint $table) {
             $table->bigIncrements('id');
@@ -17,7 +17,7 @@ return new class extends Migration {
         });
     }
 
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('webhook_calls');
     }
