@@ -5,7 +5,7 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
-    public function up()
+    public function up(): void
     {
         Schema::table('failed_jobs', function (Blueprint $table) {
             $table->uuid('uuid')->nullable()->after('id');
