@@ -1,7 +1,6 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
@@ -10,9 +9,7 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::table('media', function (Blueprint $table) {
-            $table->json('generated_conversions')->nullable();
-        });
+        Schema::rename('password_resets', 'password_reset_tokens');
     }
 
     /**
@@ -20,8 +17,6 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('media', function (Blueprint $table) {
-            //
-        });
+        Schema::rename('password_reset_tokens', 'password_resets');
     }
 };
